+++ packages/core/src/index.ts
@@ -4,3 +4,4 @@ export * from './iam/lockout';
 export * from './iam/constants';
 export * from './iam/break-glass';
 export * from './ports';
+export * from './entity';